requirements.txt
flask>=3.0
python-dotenv>=1.0
gunicorn>=22.0
requests>=2.32
Flask-Session>=0.8
Flask-WTF>=1.2
